Aula 46 - Mensagens nos cursos

+++ sistema/painel-admin/paginas/cursos/excluir.php
@@ -1,6 +1,6 @@
 <?php 
 require_once("../../../conexao.php");
-$tabela = 'categorias';
+$tabela = 'cursos';
 
 $id = $_POST['id'];
 
@@ -8,8 +8,13 @@ $query = $pdo->query("SELECT * FROM $tabela where id = '$id'");
 $res = $query->fetchAll(PDO::FETCH_ASSOC);
 $total_reg = @count($res);
 $foto = $res[0]['imagem'];
+$status = $res[0]['status'];
+if ($status == 'Aprovado'){
+	echo 'Cuidado, o curso não pode ser excluído com status de aprovado!';
+	exit();
+}
 if($foto != "sem-foto.png"){
-	@unlink('../../img/categorias/'.$foto);
+	@unlink('../../img/cursos/'.$foto);
 }
 
 $pdo->query("DELETE FROM $tabela where id = '$id'");
